Move Trip.fullExport test to travelfunds-api to get it working correctly.
This is a bit of a shortcut because I need to get an update deployed today
and I want to make sure the tests are all working.

I think this test really does belong in travelfunds-db. I'd like to
come back and refactor it to work correctly there.
